requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
dj-database-url>=2.2.0
psycopg[binary]>=3.1.18
gunicorn>=21.2.0
whitenoise>=6.7.0
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
beautifulsoup4>=4.12.0
